+++ server/sources.go
@@ -15,6 +15,7 @@ type sourceLinks struct {
 	Self        string `json:"self"`            // Self link mapping to this resource
 	Kapacitors  string `json:"kapacitors"`      // URL for kapacitors endpoint
 	Proxy       string `json:"proxy"`           // URL for proxy endpoint
+	Queries     string `json:"queries"`         // URL for the queries analysis endpoint
 	Permissions string `json:"permissions"`     // URL for all allowed permissions for this source
 	Users       string `json:"users"`           // URL for all users associated with this source
 	Roles       string `json:"roles,omitempty"` // URL for all users associated with this source
@@ -40,8 +41,9 @@ func newSourceResponse(src chronograf.Source) sourceResponse {
 		Source: src,
 		Links: sourceLinks{
 			Self:        fmt.Sprintf("%s/%d", httpAPISrcs, src.ID),
-			Proxy:       fmt.Sprintf("%s/%d/proxy", httpAPISrcs, src.ID),
 			Kapacitors:  fmt.Sprintf("%s/%d/kapacitors", httpAPISrcs, src.ID),
+			Proxy:       fmt.Sprintf("%s/%d/proxy", httpAPISrcs, src.ID),
+			Queries:     fmt.Sprintf("%s/%d/queries", httpAPISrcs, src.ID),
 			Permissions: fmt.Sprintf("%s/%d/permissions", httpAPISrcs, src.ID),
 			Users:       fmt.Sprintf("%s/%d/users", httpAPISrcs, src.ID),
 			Databases:   fmt.Sprintf("%s/%d/dbs", httpAPISrcs, src.ID),