+++ ui/src/data_explorer/components/QueryEditor.js
@@ -1,11 +1,11 @@
 import React, {PropTypes} from 'react'
-import buildInfluxQLQuery from 'utils/influxql'
 
 import DatabaseList from './DatabaseList'
 import MeasurementList from './MeasurementList'
 import FieldList from './FieldList'
 import TagList from './TagList'
 import RawQueryEditor from './RawQueryEditor'
+import buildInfluxQLQuery from 'utils/influxql'
 
 const {
   string,
@@ -15,6 +15,11 @@ const {
 
 const QueryEditor = React.createClass({
   propTypes: {
+    source: shape({
+      links: shape({
+        queries: string.isRequired,
+      }).isRequired,
+    }).isRequired,
     query: shape({
       id: string,
     }).isRequired,
@@ -31,7 +36,7 @@ const QueryEditor = React.createClass({
       toggleField: func.isRequired,
       groupByTime: func.isRequired,
       toggleTagAcceptance: func.isRequired,
-      editRawText: func.isRequired,
+      editRawTextAsync: func.isRequired,
     }).isRequired,
   },
 
@@ -68,34 +73,24 @@ const QueryEditor = React.createClass({
   },
 
   handleEditRawText(text) {
-    this.props.actions.editRawText(this.props.query.id, text)
+    const {source: {links}, query} = this.props
+    this.props.actions.editRawTextAsync(links.queries, query.id, text)
   },
 
   render() {
+    const {query, timeRange} = this.props
+    const q = query.rawText || buildInfluxQLQuery(timeRange, query) || ''
+
     return (
       <div className="query-builder--tab-contents">
         <div>
-          {this.renderQuery()}
+          <RawQueryEditor query={q} config={query} onUpdate={this.handleEditRawText} />
           {this.renderLists()}
         </div>
       </div>
     )
   },
 
-  renderQuery() {
-    const {query, timeRange} = this.props
-    const statement = query.rawText || buildInfluxQLQuery(timeRange, query) || 'Select a database, measurement, and field below.'
-
-    if (typeof query.rawText !== 'string') {
-      return (
-        <div className="query-builder--query-preview">
-          <pre><code>{statement}</code></pre>
-        </div>
-      )
-    }
-
-    return <RawQueryEditor query={query} onUpdate={this.handleEditRawText} />
-  },
 
   renderLists() {
     const {query} = this.props